pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "colorama>=0.4.6",
    "dnspython>=2.7.0",
    "ipwhois>=1.3.0",
    "python-nmap>=0.7.1",
    "python-whois>=0.9.5",
    "requests>=2.32.5",
]
